Allow bootstrap path to use new format image
There is a place in the daemon where the image can get updated as part
of the bootstrap flow. For completeness, we should make sure that it can
use a "new format" image, if it were to ever get one.

This makes the bootstrap flow check to see if the image it got is a new
format image, so it can apply it properly.

+++ pkg/daemon/daemon.go
@@ -1410,16 +1410,30 @@ func (dn *Daemon) checkStateOnFirstRun() error {
 		osMatch := dn.checkOS(targetOSImageURL)
 		if !osMatch {
 			glog.Infof("Bootstrap pivot required to: %s", targetOSImageURL)
-			// This only returns on error
-			osImageContentDir, err := ExtractOSImage(targetOSImageURL)
+
+			// Check to see if we have a layered/new format image
+			isLayeredImage, err := dn.NodeUpdaterClient.IsBootableImage(targetOSImageURL)
 			if err != nil {
-				return err
+				return fmt.Errorf("Error checking type of target image: %w", err)
 			}
-			if err := updateOS(state.currentConfig, osImageContentDir); err != nil {
-				return err
-			}
-			if err := os.RemoveAll(osImageContentDir); err != nil {
-				return err
+
+			if isLayeredImage {
+				// If this is a new format image, we don't have to extract it,
+				// we can just update it the proper way
+				if err := updateLayeredOS(state.currentConfig); err != nil {
+					return err
+				}
+			} else {
+				osImageContentDir, err := ExtractOSImage(targetOSImageURL)
+				if err != nil {
+					return err
+				}
+				if err := updateOS(state.currentConfig, osImageContentDir); err != nil {
+					return err
+				}
+				if err := os.RemoveAll(osImageContentDir); err != nil {
+					return err
+				}
 			}
 			if err := dn.finalizeBeforeReboot(state.currentConfig); err != nil {
 				return err
